Enable integer overflow sanitizer in tests.
Some vendors compile Protobuf clients with unsigned integer
overflow sanitizer. This sanitizer is not supported upstream,
because unsigned overflow is not undefined behavior. We carry
local modifications to support these vendors, but so far we
haven't had test coverage for this use case.

Enable the sanitizer in tests and add locally disable it in
test code that intentionally uses unsigned overflow to prevent
any future regressions.

Bug: 452854573
Test: presubmit
Change-Id: I0b0561a5d06663088617b578d85dfc4e87370c9a